requirements.txt
Django==3.2.16
djangorestframework==3.14.0
# psycopg2-binary==2.9.5
djoser==2.1.0
django-cors-headers==3.14.0
python-dotenv==1.0.0
Pillow==9.5.0
celery[redis]==5.2.7
flower==1.2.0
